fine tune user permissions

+++ papermerge/core/views/tokens.py
@@ -4,6 +4,8 @@ import logging
 from django.contrib.auth.decorators import login_required
 from django.contrib import messages
 from django.shortcuts import render, redirect
+from django.utils.translation import gettext as _
+
 from knox.models import AuthToken
 from papermerge.core.forms import AuthTokenForm
 
@@ -18,9 +20,18 @@ def tokens_view(request):
         go_action = request.POST['action']
 
         if go_action == 'delete_selected':
-            AuthToken.objects.filter(
-                digest__in=selected_action
-            ).delete()
+            if request.user.has_perm('delete_authtoken'):
+                AuthToken.objects.filter(
+                    digest__in=selected_action
+                ).delete()
+            else:
+                messages.info(
+                    request,
+                    _("You don't have %(model)s %(delete)s permissions") % {
+                        'delete': _('delete'),
+                        'model': _('token')
+                    }
+                )
 
     tokens = AuthToken.objects.all()
 
@@ -29,6 +40,7 @@ def tokens_view(request):
         'admin/tokens.html',
         {
             'tokens': tokens,
+            'has_perm_add_authtoken': request.user.has_perm('add_authtoken')
         }
     )
 
@@ -37,22 +49,30 @@ def tokens_view(request):
 def token_view(request):
 
     if request.method == 'POST':
+        if request.user.has_perm('add_authtoken'):
+            form = AuthTokenForm(request.POST)
 
-        form = AuthTokenForm(request.POST)
+            if form.is_valid():
+                instance, token = AuthToken.objects.create(
+                    user=request.user,
+                    expiry=timedelta(hours=form.cleaned_data['hours'])
+                )
+                html_message = "Please remember the token: "
+                html_message += f" <span class='text-success'>{token}</span>"
+                html_message += " It won't be displayed again."
+                messages.success(
+                    request, html_message
+                )
 
-        if form.is_valid():
-
-            instance, token = AuthToken.objects.create(
-                user=request.user,
-                expiry=timedelta(hours=form.cleaned_data['hours'])
+                return redirect('core:tokens')
+        else:
+            messages.info(
+                request,
+                _("You don't have %(model)s %(add)s permissions") % {
+                    'add': _('add'),
+                    'model': _('token')
+                }
             )
-            html_message = "Please remember the token: "
-            html_message += f" <span class='text-success'>{token}</span>"
-            html_message += " It won't be displayed again."
-            messages.success(
-                request, html_message
-            )
-            return redirect('core:tokens')
 
     form = AuthTokenForm()
 
@@ -61,5 +81,6 @@ def token_view(request):
         'admin/token.html',
         {
             'form': form,
+            'has_perm_add_authtoken': request.user.has_perm('add_authtoken')
         }
     )
